salesintel-ai/backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.32
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
pgvector>=0.3.5
alembic>=1.13.0

# Redis / Celery
redis>=5.0.0
celery[redis]>=5.4.0

# OpenAI
openai>=1.40.0

# Vector DB
qdrant-client>=1.11.0

# Elasticsearch
elasticsearch[async]>=8.12.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
pydantic[email]>=2.8.0

# Scraping
httpx>=0.27.0
playwright>=1.45.0
playwright-stealth>=1.0.6
trafilatura>=1.12.0
beautifulsoup4>=4.12.3
lxml>=5.2.2

# RAG / Embeddings
sentence-transformers>=3.0.0
fastembed>=0.2.6

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0

# Utilities
tenacity>=8.5.0
